Move startpk-test-api into startpk
