+++ startpk/include/beetree/test-api/mocks/mock-navigation-service.hpp
@@ -1,5 +1,5 @@
 #include "beetree/services/i-navigation-service.hpp"
-#include "beetree/test/gtest-ns.hpp"
+#include "beetree/test-api/gtest-ns.hpp"
 
 namespace bte {
 namespace test {
